# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "The clinic's API",
        "version": "1"
    },
    "servers": [
        {
            "url": "https://localhost",
            "description": "Optional server description, e.g. Main (production) server"
        }
    ],
...
    "paths": {
        "/isAuthenticated": {
            "get": {
                "operationId": "is-authenticated",
                "tags": [
                    "General"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "authenticated": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
...
        "/locales": {
            "get": {
                "description": "Gets all the locale options available.",
                "tags": [
                    "General"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "longName": {
                                                "type": "string",
                                                "example": "English"
                                            },
                                            "shortName": {
                                                "type": "string",
                                                "example": "en"
                                            },
                                            "direction": {
                                                "type": "string",
                                                "description": "The direction in which the language is written.",
                                                "example": "ltr"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "",
                "tags": [
                    "Accounts"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object",
                                        "required": [
                                            "password",
                                            "username"
                                        ],
                                        "properties": {
                                            "username": {
                                                "type": "string",
                                                "example": "fooname",
                                                "format": "username"
                                            },
                                            "password": {
                                                "type": "string",
                                                "format": "password"
                                            },
                                            "remember": {
                                                "type": "boolean"
                                            }
                                        }
                                    },
                                    {
                                        "type": "object",
                                        "required": [
                                            "password",
                                            "email"
                                        ],
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "example@example.example",
                                                "format": "email"
                                            },
                                            "password": {
                                                "type": "string",
                                                "format": "password"
                                            },
                                            "remember": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Successful, redirects the user to the home page."
                    },
                    "400": {
                        "$ref": "components/responses/400/Accounts/login.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
...
        "/isEmailVerified": {
            "get": {
                "description": "",
                "tags": [
                    "Accounts"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "applicaiton/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "verified": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
...
        "/logout": {
            "get": {
                "description": "",
                "tags": [
                    "Accounts"
                ],
                "responses": {
                    "301": {
                        "description": "redirects user to '/' or to previous url"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
...
        "/account/{roleName}": {
            "post": {
                "description": "Creats a new user based on the provided roleName path variable.",
                "tags": [
                    "Accounts"
                ],
                "parameters": [
                    {
                        "name": "roleName",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "phonenumber",
                                    "firstname",
                                    "lastname",
                                    "username",
                                    "password",
                                    "password_confirmation",
                                    "gender"
                                ],
                                "properties": {
                                    "phonenumber": {
                                        "$ref": "components/schemas/data-types.json#/phonenumber"
                                    },
                                    "firstname": {
                                        "$ref": "components/schemas/data-types.json#/firstname"
                                    },
                                    "lastname": {
                                        "$ref": "components/schemas/data-types.json#/lastname"
                                    },
                                    "username": {
                                        "$ref": "components/schemas/data-types.json#/username"
                                    },
                                    "email": {
                                        "$ref": "components/schemas/data-types.json#/email"
                                    },
                                    "password": {
                                        "$ref": "components/schemas/data-types.json#/password"
                                    },
                                    "password_confirmation": {
                                        "$ref": "components/schemas/data-types.json#/password_confirmation"
                                    },
                                    "gender": {
                                        "$ref": "components/schemas/data-types.json#/gender"
                                    },
                                    "avatar": {
                                        "$ref": "components/schemas/data-types.json#/avatar"
                                    }
                                },
                                "additionalProperties": true
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "user"
                                    ],
                                    "properties": {
                                        "user": {
                                            "$ref": "components/schemas/Accounts/user.json"
                                        }
                                    },
                                    "additionalProperties": true
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "components/responses/400/Accounts/accounts-store.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
        "/account/{placeholder}": {
            "get": {
                "description": "Retrieves a user account based on a placeholder",
                "tags": [
                    "Accounts"
                ],
                "parameters": [
                    {
                        "name": "placeholder",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "components/schemas/data-types.json#/username"
                                },
                                {
                                    "$ref": "components/schemas/data-types.json#/email"
                                },
                                {
                                    "$ref": "components/schemas/data-types.json#/phonenumber"
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "user"
                                    ],
                                    "properties": {
                                        "user": {
                                            "$ref": "components/schemas/Accounts/user.json"
                                        }
                                    },
                                    "additionalProperties": true
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "components/responses/400/Accounts/accounts-show.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
        "/account": {
            "get": {
                "description": "Retrieves the authenticated user's account.",
                "tags": [
                    "Accounts"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "user"
                                    ],
                                    "properties": {
                                        "user": {
                                            "$ref": "components/schemas/Accounts/user.json"
                                        }
                                    },
                                    "additionalProperties": true
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "components/responses/400/Accounts/accounts-show.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
        "/account/{accountId}": {
            "put": {
                "description": "Updates a user's account.",
                "tags": [
                    "Accounts"
                ],
                "parameters": [
                    {
                        "name": "accountId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "components/schemas/data-types.json#/id"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstname": {
                                        "$ref": "components/schemas/data-types.json#/firstname"
                                    },
                                    "lastname": {
                                        "$ref": "components/schemas/data-types.json#/lastname"
                                    },
                                    "username": {
                                        "$ref": "components/schemas/data-types.json#/username"
                                    },
                                    "gender": {
                                        "$ref": "components/schemas/data-types.json#/gender"
                                    },
                                    "avatar": {
                                        "$ref": "components/schemas/data-types.json#/avatar"
                                    }
                                },
                                "additionalProperties": true
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "The updated user",
                                    "type": "object",
                                    "required": [
                                        "user"
                                    ],
                                    "properties": {
                                        "user": {
                                            "$ref": "components/schemas/Accounts/user.json"
                                        }
                                    },
                                    "additionalProperties": true
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "components/responses/400/Accounts/accounts-show.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            },
            "delete": {
                "description": "Updates a user's account.",
                "tags": [
                    "Accounts"
                ],
                "parameters": [
                    {
                        "name": "accountId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "components/schemas/data-types.json#/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "message"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "components/responses/400/Accounts/accounts-show.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
...
        "/role": {
            "get": {
                "description": "",
                "tags": [
                    "Roles"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
...
        "/settings": {
            "get": {
                "description": "",
                "tags": [
                    "BusinessDefault"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
...
        "/visit/laser/check": {
            "post": {
                "description": "",
                "tags": [
                    "Visits"
                ],
                "parameters": [
                    {
                        "name": "businessName",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/paths/~1visit~1laser/post/requestBody"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "components/responses/400/Visits/visits-laser.json"
                    },
                    "500": {
                        "$ref": "components/responses/500.json#/not-empty"
                    }
                }
            }
        },
...
    }
}